+++ debug/plugins/prism-es6/components/prism-elixir.js
@@ -717,6 +717,7 @@ Prism.languages.elixir = {
 
 
 
+
       // See interpolation below
     } }, { pattern: /("""|''')[\s\S]*?\1/, greedy: true, inside: {