+++ install.sh
@@ -104,7 +104,7 @@ module.exports = {
             ,"@flexvue/recyclerview":`./${debugPath}/flexvue/core/recyclerview.class.js`
             ,"@flexvue/caches":`./${debugPath}/flexvue/core/caches.class.js`
             ,"@flexvue/elements/types":`./${debugPath}/flexvue/components/ElementsTypes.interface.js`
-            ,"@flexvue/elements/html":`./${debugPath}/flexvue/components/ElementsHtml.class.js`
+            ,"@flexvue/elements/manager":`./${debugPath}/flexvue/components/ElementsManager.class.js`
             ,"@flexvue/elements/attrs":`./${debugPath}/flexvue/components/ElementsAttributeOptions.class.js`
             ,"@flexvue/elements/create":`./${debugPath}/flexvue/components/ElementsCreate.class.js`
             ,"@editorjs/editorjs": `./${debugPath}/plugins/@editorjs/editorjs/dist/editorjs.js`
@@ -181,7 +181,7 @@ echo '{
           "@flexvue/recyclerview": ["src/flexvue/core/recyclerview.class.ts"],
           "@flexvue/caches": ["src/flexvue/core/caches.class.ts"],
           "@flexvue/elements/types": ["src/flexvue/components/ElementsTypes.interface.ts"],
-          "@flexvue/elements/html": ["src/flexvue/components/ElementsHtml.class.ts"],
+          "@flexvue/elements/manager": ["src/flexvue/components/ElementsManager.class.ts"],
           "@flexvue/elements/attrs": ["src/flexvue/components/ElementsAttributeOptions.class.ts"],
           "@flexvue/elements/create": ["src/flexvue/components/ElementsCreate.class.ts"],
           "@editorjs/editorjs": ["node_modules/@editorjs/editorjs"],